chore: make xstate version default
